+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - change hasSuperUser to DomainType
 - Open domain's admin is changed from zero address to normal address
 - Recon configuration
+- Recon register account handler
 
 ## v0.2.5